Fix removePlayerFromGroup race condition
- converted unit tests to use real redis instance

+++ socketServer/src/controllers/GroupController.js
@@ -44,7 +44,7 @@ module.exports = class GroupController extends BaseController {
         });
 
         this.on(serverEvents.groupInviteDecline, (data) => {
-            return groupService.removePlayerFromGroupPending(this.battleNetId, data.eventData, this.socket, this.namespace);
+            return groupService.declineGroupInvite(this.battleNetId, data.eventData, this.socket, this.namespace);
         });
 
         this.on(serverEvents.groupInviteCancel, (data) => {
@@ -56,7 +56,9 @@ module.exports = class GroupController extends BaseController {
         });
 
         this.on(serverEvents.disconnect, () => {
-            return groupService.removePlayerFromGroup(this.battleNetId, this.groupId, this.socket, this.namespace);
+            if (this.groupId) {
+                return groupService.removePlayerFromGroup(this.battleNetId, this.groupId, this.socket, this.namespace);
+            }
         });
     }
 };
